third_party/blink/web_tests/flag-specific/highdpi/fast/css-grid-layout/grid-auto-repeat-huge-grid-002-expected.txt

This is a testharness.js-based test.
FAIL Test that we don't get more than kGridMaxTracks repetitions even on very wide grids with gaps. assert_equals: expected 30 but got 20
FAIL Test that we don't get more than kGridMaxTracks repetitions even on very tall grids with gaps. assert_equals: expected 30 but got 20
Harness: the test ran to completion.
